Issue #146: Move creation of test environments into its own Node.js module
